# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the backend API for the HS project",
        "title": "HS Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/api/facets": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns facets",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Facets"
                ],
                "summary": "Get facets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.GetFacetResponseItem"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new facet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Facets"
                ],
                "summary": "Create facet",
                "parameters": [
                    {
                        "description": "Facet",
                        "name": "facet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateFacetInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateFacetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/facets/{facetId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a facet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Facets"
                ],
                "summary": "Delete facet",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Facet ID",
                        "name": "facetId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates a facet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Facets"
                ],
                "summary": "Update facet",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Facet ID",
                        "name": "facetId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Facet",
                        "name": "facet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateFacetInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateFacetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/prisms/{prismId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a prism",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prisms"
                ],
                "summary": "Delete prism",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Prism ID",
                        "name": "prismId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates a prism",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prisms"
                ],
                "summary": "Update prism",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Prism ID",
                        "name": "prismId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Prism",
                        "name": "prism",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdatePrismInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdatePrismResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/prisms/{prismId}/activate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Activates a prism",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prisms"
                ],
                "summary": "Activate prism",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Prism ID",
                        "name": "prismId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.CreateFacetEnrichedConfig": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CreateFacetEnrichedConfigItem"
                    }
                }
            }
        },
        "dto.CreateFacetEnrichedConfigItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "service": {
                    "$ref": "#/definitions/enum.CommunicationService"
                },
                "status": {
                    "$ref": "#/definitions/enum.FacetStatus"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "dto.CreateFacetInput": {
            "type": "object",
            "required": [
                "color",
                "configuration",
                "privateLabel",
                "publicLabel"
            ],
            "properties": {
                "color": {
                    "type": "string"
                },
                "configuration": {
                    "$ref": "#/definitions/json.FacetConfig"
                },
                "privateLabel": {
                    "type": "string"
                },
                "publicLabel": {
                    "type": "string"
                }
            }
        },
        "dto.CreateFacetResponse": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "configuration": {
                    "$ref": "#/definitions/dto.CreateFacetEnrichedConfig"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "privateLabel": {
                    "type": "string"
                },
                "publicLabel": {
                    "type": "string"
                }
            }
        },
...
        "dto.GetFacetEnrichedConfig": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GetFacetEnrichedConfigItem"
                    }
                }
            }
        },
        "dto.GetFacetEnrichedConfigItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "service": {
                    "$ref": "#/definitions/enum.CommunicationService"
                },
                "status": {
                    "$ref": "#/definitions/enum.FacetStatus"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "dto.GetFacetResponseItem": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "configuration": {
                    "$ref": "#/definitions/dto.GetFacetEnrichedConfig"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "privateLabel": {
                    "type": "string"
                },
                "publicLabel": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateFacetEnrichedConfig": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.UpdateFacetEnrichedConfigItem"
                    }
                }
            }
        },
        "dto.UpdateFacetEnrichedConfigItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "service": {
                    "$ref": "#/definitions/enum.CommunicationService"
                },
                "status": {
                    "$ref": "#/definitions/enum.FacetStatus"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateFacetInput": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "configuration": {
                    "$ref": "#/definitions/json.FacetConfig"
                },
                "privateLabel": {
                    "type": "string"
                },
                "publicLabel": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateFacetResponse": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "configuration": {
                    "$ref": "#/definitions/dto.UpdateFacetEnrichedConfig"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "privateLabel": {
                    "type": "string"
                },
                "publicLabel": {
                    "type": "string"
                }
            }
        },
        "dto.UpdatePrismEnrichedConfig": {
            "type": "object",
            "properties": {
                "base": {
                    "$ref": "#/definitions/dto.UpdatePrismFacet"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.UpdatePrismEnrichedUserItem"
                    }
                }
            }
        },
        "dto.UpdatePrismEnrichedUserItem": {
            "type": "object",
            "properties": {
                "facet": {
                    "$ref": "#/definitions/dto.UpdatePrismFacet"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "dto.UpdatePrismFacet": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "privateLabel": {
                    "type": "string"
                },
                "publicLabel": {
                    "type": "string"
                }
            }
        },
        "dto.UpdatePrismInput": {
            "type": "object",
            "properties": {
                "configuration": {
                    "$ref": "#/definitions/json.PrismConfig"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.UpdatePrismResponse": {
            "type": "object",
            "properties": {
                "configuration": {
                    "$ref": "#/definitions/dto.UpdatePrismEnrichedConfig"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "enum.CommunicationService": {
            "type": "string",
            "enum": [
                "faceToFace",
                "phone",
                "message",
                "email",
                "discord",
                "microsoftTeams"
            ],
            "x-enum-varnames": [
                "FaceToFace",
                "Phone",
                "Message",
                "Email",
                "Discord",
                "MicrosoftTeams"
            ]
        },
        "enum.FacetStatus": {
            "type": "string",
            "enum": [
                "available",
                "emergencyOnly"
            ],
            "x-enum-varnames": [
                "Available",
                "EmergencyOnly"
            ]
        },
        "error.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "json.FacetConfig": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/json.FacetConfigItem"
                    }
                }
            }
        },
        "json.FacetConfigItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/enum.FacetStatus"
                }
            }
        },
        "json.PrismConfig": {
            "type": "object",
            "properties": {
                "base": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/json.PrismConfigUserItem"
                    }
                }
            }
        },
        "json.PrismConfigUserItem": {
            "type": "object",
            "properties": {
                "facetId": {
                    "type": "integer"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and your token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}